projects/gliner2-poc/pyproject.toml
[project]
name = "gliner2-poc"
version = "0.1.0"
description = "GLiNER2 POC evaluation: unified NER + classification vs specialized baselines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gliner2>=1.1.0",
    "transformers>=4.40.0",
    "datasets>=2.19.0",
    "tokenizers>=0.19.0",
    "torch>=2.2.0",
    "spacy>=3.7.0",
    "seqeval>=1.2.2",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "notebook>=7.0.0",
    "accelerate>=0.28.0",
    "huggingface-hub>=0.22.0",
    "gliner>=0.2.19",
    "flair>=0.14.0",
    "setfit>=1.0.0",
    "fasttext-wheel>=0.9.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
